Don't prompt for scans folder during Save As
Even if AutoImg is on, it should only prompt when a new
file is loaded, not when a file is saved.

Fixes #812
